Fix warning in trg

+++ trg/gdl/modules/trggdlDQM/src/TRGGDLDQMModule.cc
@@ -455,7 +455,7 @@ void TRGGDLDQMModule::event()
     int psn_tmp[3] = {0};
     int ftd_tmp[3] = {0};
     int itd_tmp[5] = {0};
-    for (unsigned j = 0; j < nword_input; j++) {
+    for (unsigned j = 0; j < (unsigned)nword_input; j++) {
       itd_tmp[j] = h_0->GetBinContent(clk, 1 + ee_itd[j]);
       itd[j] |= itd_tmp[j];
       for (int i = 0; i < 32; i++) {
@@ -480,7 +480,7 @@ void TRGGDLDQMModule::event()
         if (ftd_tmp[1] & (1 << i)) h_f->SetBinContent(clk, i + 1 + 32, 1);
       }
     } else {
-      for (unsigned j = 0; j < nword_output; j++) {
+      for (unsigned j = 0; j < (unsigned)nword_output; j++) {
         psn_tmp[j] = h_0->GetBinContent(clk, 1 + ee_psn[j]);
         ftd_tmp[j] = h_0->GetBinContent(clk, 1 + ee_ftd[j]);
         psn[j] |= psn_tmp[j];
@@ -507,10 +507,10 @@ void TRGGDLDQMModule::event()
   h_ftd->Fill(-0.5);
   h_psn->Fill(-0.5);
   for (int i = 0; i < 32; i++) {
-    for (unsigned j = 0; j < nword_input; j++) {
+    for (unsigned j = 0; j < (unsigned)nword_input; j++) {
       if (itd[j] & (1 << i)) h_itd->Fill(i + 0.5 + 32 * j);
     }
-    for (unsigned j = 0; j < nword_output; j++) {
+    for (unsigned j = 0; j < (unsigned)nword_output; j++) {
       if (ftd[j] & (1 << i)) h_ftd->Fill(i + 0.5 + 32 * j);
       if (psn[j] & (1 << i)) h_psn->Fill(i + 0.5 + 32 * j);
     }
